+++ apworld/tf2_mvm/__init__.py
@@ -231,9 +231,7 @@ class TF2MvMWorld(World):
         return [
             mission
             for mission in data.MISSIONS
-            if mission.playable
-            and mission.difficulty in allowed
-            and mission.name not in excluded
+            if mission.playable and mission.difficulty in allowed and mission.name not in excluded
         ]
 
     @staticmethod